+++ Customer.Inquiries.DataAccess/Models/Transaction.cs
@@ -9,7 +9,7 @@ namespace Customer.Inquiries.DataAccess.Models
     public class Transaction : BaseEntity
     {
         [Key]
-        public int TransactionId { get; set; }
+        public long TransactionId { get; set; }
 
         public DateTimeOffset TransactionDateTime { get; set; }
 
@@ -23,7 +23,7 @@ namespace Customer.Inquiries.DataAccess.Models
         public ETransactionStatus Status { get; set; }
 
         [ForeignKey("Customer")]
-        public int CustomerId { get; set; }
+        public long CustomerId { get; set; }
 
         [ForeignKey("CustomerId")]
         public virtual Customer Customer { get; set; }
